Add files via upload
This program listens for DNS queries for "www.example.com" from a host with IP 10.9.0.53 on the interface 'br-9bff7edb888c'. When it captures such a query, it creates a DNS Response, falsely suggesting that it comes from the queried DNS Server.
The response contains:

An Answer Section that asserts 'www.example.com' resolves to the IP address '1.1.1.1'.
An Authority Section claiming that 'example.com' is served by two nameservers: 'ns.attacker32.com' and 'ns.example.com'.
An Additional Section that provides IP addresses for 'ns.attacker32.com', 'ns.example.net', and 'www.facebook.com'. These IP addresses are all fake and the protocol does not require the request.

After constructing the response, the script sends it to the querying host (10.9.0.53) in order to trick it into accepting the fraudulent mappings. This type of attack is known as DNS Spoofing or DNS Cache Poisoning attack.
The idea behind such attacks is to divert traffic away from legitimate servers and towards malicious ones.